separate components by pages and components
