+++ src/pages/NewGame.css
